# shop/api_specs/api_spec.json
{
    "swagger": "2.0",
    "host": "127.0.0.1:8000",
    "basePath": "/api/shop/",
    "info": {
        "version": "1.0.0",
        "title": "Simple API",
        "description": "A simple API to learn how to write OpenAPI Specification"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "security": [],
    "securityDefinitions": {
      "oauth": {
          "tokenUrl": "http://auth.ibtspl.com/oauth2/",
          "flow": "password",
          "scopes": {
              "read": "read users",
              "write": "create users",
              "update": "update users",
              "delete": "delete users",
              "superuser": "super user permission" 
          },
          "type": "oauth2"
      }
  },
...
 "parameters": {
...
        "ProductId": {
            "description": "Product id",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
        },
...
    },
...
    "paths": {
...
          "/product/update/{id}/":{
              "put":{
                  "parameters": [
                  {
                      "$ref": "#/parameters/ProductId"
                  }
              ],
              "summary": "Updates a Product",
              "description": "Updates a Product.",
              "operationId": "update_product",
              "security": [
                    {
                        "oauth": ["update"]
                    }
                ],
              "responses": {
                  "200": {
                      "description":"A product updated Successfully.....!"
                  }
              }
          }
          },
...
}
}